requirements.txt
streamlit
openai
fpdf
pdfplumber
python-docx
